requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.6b0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
customerio==1.2
defusedxml==0.7.1
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
google-api-core==1.31.2
google-api-python-client==2.15.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.1
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
model-mommy==2.0.0
mypy-extensions==0.4.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
pathspec==0.8.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
